Praticamente tudo

+++ src/usuarios/usuario.service.ts
@@ -1,8 +1,11 @@
+import { USUARIOS } from '../common/usuario.const';
 import { Usuario } from './interfaces/usuario.interface';
 import { UsuarioRepository } from './usuario.repository';
 
 export class UsuarioService {
-  constructor(private readonly usuarioRepository: UsuarioRepository) {}
+  constructor(private readonly usuarioRepository: UsuarioRepository) {
+    this.usuarioRepository = new UsuarioRepository();
+  }
 
   async getUsuarioById(id: string) {
     return this.usuarioRepository.getUsuarioById(id);
@@ -17,14 +20,24 @@ export class UsuarioService {
   }
 
   async createUsuario(usuario: Usuario) {
+    if (USUARIOS.find((u) => u.email === usuario.email))
+      throw new Error('Email already in use');
+    if (!usuario.nome || !usuario.email || !usuario.senha)
+      throw new Error('Invalid fields');
     return this.usuarioRepository.createUsuario(usuario);
   }
 
   async updateUsuario(id: string, usuario: Usuario) {
+    if (USUARIOS.find((u) => u.email === usuario.email))
+      throw new Error('Email already in use');
+    if (!usuario.nome || !usuario.email || !usuario.senha)
+      throw new Error('Invalid fields');
+    if (!USUARIOS.find((u) => u.id === id)) throw new Error('User not found');
     return this.usuarioRepository.updateUsuario(id, usuario);
   }
 
   async deleteUsuario(id: string) {
+    if (!USUARIOS.find((u) => u.id === id)) throw new Error('User not found');
     return this.usuarioRepository.deleteUsuario(id);
   }
 }
